requirements.txt
pytorch==1.12.0
torchvision==0.13.0
tensorboard==2.9.1
numpy==1.22.3
pandas==1.4.3
matplotlib==3.5.1
tqdm==4.64.0
pyyaml==6.0
gpytorch==1.8.0
scikit-learn==1.0.2